+++ src/MelloSilveiraTools.MechanicsOfMaterials.Optimizations/Services/ExperimentalData/ExperimentalDataSettings.cs
@@ -1,6 +1,0 @@
-﻿namespace MelloSilveiraTools.MechanicsOfMaterials.Optimizations.Services.ExperimentalData;
-
-public record ExperimentalDataSettings
-{
-    public int FileWriterBoundedCapacity { get; init; } = 10000;
-}